extensions to docs
