+++ app/src/docs/Extensions/index.tsx
@@ -1,0 +1,10 @@
+/* @license Copyright 2023 @paritytech/polkadot-cloud authors & contributors
+SPDX-License-Identifier: GPL-3.0-only */
+
+import Doc from "./index.mdx";
+
+export const Extensions = () => (
+  <div className="doc">
+    <Doc npm="@polkadot-cloud/community" />
+  </div>
+);
