--- app/src/docs/extensions.mdx
@@ -1,7 +1,0 @@
-import { NPM } from "./lib/NPM";
-
-# Extensions
-
-## Web Extension icons in SVG and JSX format.
-
-<NPM npm={props.npm} />
