+++ app/src/styles/app.scss
@@ -182,6 +182,12 @@ SPDX-License-Identifier: GPL-3.0-only */
       margin-bottom: 2rem;
     }
   }
+
+  h4 {
+    color: var(--text-color-primary);
+    font-size: 1.25rem;
+    margin-bottom: 1.5rem;
+  }
   
   .demo {
     border: 1px solid var(--border-secondary-color);
@@ -199,20 +205,26 @@ SPDX-License-Identifier: GPL-3.0-only */
   margin-bottom: 1.25rem;
 }
 
-.svg-box {
-  color: var(--text-color-primary);
-  width: 7.5rem;
-  height: 7.5rem;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  margin: 0 2rem 1rem 0;
 
-  > svg {
-    max-width: 100%;
-    max-height: 100%;
+  .svg-box {
+    color: var(--text-color-primary);
+    width: 5rem;
+    height: 5rem;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    margin: 0 3rem 0rem 0;
+
+    > svg {
+      max-width: 100%;
+      max-height: 100%;
+    }
+
+    &:last-child {
+      margin-right: 0;
+    }
   }
-}
+
 
 .code {
   overflow: hidden;
